+++ CHANGELOG.md
@@ -55,3 +55,5 @@
 
 - Added Opera Stations.
 - Added an animation to the LIVE flag
+- Added Latino Stations.
+- Fixed Accordion height issue.